Refactor Interface names and locations to be consistent
